fetch-service/berlin_covid_prediction/requirements.txt
pycouchdb
python-dotenv
urllib3
pandas
numpy
